+++ src/cadqueryhelper/Hinge.py
@@ -1,0 +1,144 @@
+import cadquery as cq
+from . import Base
+
+class Hinge(Base):
+    def __init__(self):
+        super().__init__()
+        #parameters
+        self.length = 40
+        self.radius = 2
+        self.segments = 4 
+        self.pad = 1
+
+        self.base_inset = 0.6
+        self.key_length = 1.5
+        self.key_width = 0.5
+
+        self.tab_length = 10
+        self.rotate_deg = 0
+        self.plate_spacer = 0.4
+
+        #parts
+        self.h_parts = None
+        self.join_plate = None
+        
+    def __make_driver(self, length):
+        points = [
+            (0,0),
+            (length,0),
+            (length,self.base_inset),
+            (length+self.key_length, self.radius-self.key_width),
+            (length+self.key_length, self.radius),
+            (0-self.key_length, self.radius),
+            (0-self.key_length, self.radius-self.key_width),
+            (0,self.base_inset)
+        ]
+        
+        cyl = (
+            cq.Workplane("XY")
+            .center(0,0)
+            .polyline(points).close()
+            .translate((-1*((length)/2),-self.radius,0))
+            .revolve(360,(0,self.radius,0),(1,self.radius,0))
+        ).translate((-length/2,-self.radius,0))
+        
+        return cyl
+    
+    def __make_receiver(self,length):
+        points = [
+            (0,0),
+            (length,0),
+            (length,self.base_inset),
+            (length-self.key_length, self.radius-self.key_width),
+            (length-self.key_length, self.radius),
+            (0+self.key_length, self.radius),
+            (0+self.key_length, self.radius-self.key_width),
+            (0,self.base_inset)
+        ]
+        
+        cyl = (
+            cq.Workplane("XY")
+            .center(0,0)
+            .polyline(points).close()
+            .translate((-1*((length)/2),-self.radius,0))
+            .revolve(360,(0,self.radius,0),(1,self.radius,0))
+        ).translate((-length/2,-self.radius,0))
+        return cyl
+    
+    def __make_tab(self, length, side):
+        direction = -1
+        if side:
+            direction = 1
+            
+        cut_cylinder = (
+            cq.Workplane("XY")
+            .cylinder(length, self.radius-.1).rotate((0,1,0),(0,0,0),90)
+        )
+            
+        tab = (
+            cq.Workplane("XY")
+            .box(length,self.tab_length,self.radius)
+        ).translate((0,direction*(self.tab_length/2),-self.radius/2))
+        return tab.cut(cut_cylinder)
+        
+    def __hinge_cylinder(self, length, side):
+        if side:
+            cyl = self.__make_driver(length)
+        else:
+            cyl = self.__make_receiver(length)
+        
+        tab = self.__make_tab(length, side)
+        combined = cyl.union(tab)
+        
+        if side:
+            combined = combined.rotate((1,0,0),(0,0,0),self.rotate_deg)
+        return combined
+
+
+    def __make_segments(self):
+        segment_length = self.length/self.segments
+        h_parts = cq.Workplane("XY")
+        for i in range(0,self.segments):
+            cyl = self.__hinge_cylinder(
+                length = segment_length-self.pad/2,
+                side = i%2==0
+            )
+        
+            if i == self.segments -1:
+                #log(f"end {segment_length-pad/2}")
+                cyl = cyl.translate((self.pad/4,0,0))
+            elif i == 0:
+                #log(f"begin {segment_length-pad/2}")
+                cyl = cyl.translate((-self.pad/4,0,0))
+            h_parts = h_parts.add(cyl.translate(((segment_length)*i,0,0)))
+            
+        h_parts = h_parts.translate((segment_length/2-self.length/2,0,0))
+        self.h_parts = h_parts
+        
+    def __make_join_plate(self):
+        join_plate = (
+            cq.Workplane("XY")
+            .box(
+                self.length,
+                self.tab_length-self.radius-self.plate_spacer, 
+                self.radius
+            )
+            .translate((0,0,-self.radius/2))
+        )
+        self.join_plate = join_plate
+        
+    def make(self, parent=None):
+        super().make(parent)
+        self.__make_segments()
+        self.__make_join_plate()
+
+    def build(self):
+        super().build()
+        p_x_translate = (self.tab_length/2)+(self.radius+self.plate_spacer)/2
+        hinge_build = (
+            cq.Workplane("XY")
+            .union(self.h_parts)
+            .union(self.join_plate.translate((0,p_x_translate,0)).rotate((1,0,0),(0,0,0),self.rotate_deg))
+            .union(self.join_plate.translate((0,-1*p_x_translate,0)))
+        )
+        return hinge_build