+++ src/cadqueryhelper/__init__.py
@@ -1,2 +1,4 @@
 from .series import series
 from .irregularGrid import irregular_grid
+from .Base import Base
+from .Hinge import Hinge